fix(semantic-link): type=25 列级联删除与锚点文本恢复
补齐 NoteColumnServiceImpl.deleteNoteColumnByIds 中 type=25 语义关联列
分支的级联清理与文本恢复，修复删除后残留孤儿 NoteNotelink 记录与笔记死链。

- U1: 新增 NoteBlockContentService，用 Jsoup 递归遍历 NoteBlock.property
  JSON 并替换匹配的 a[data-type=semantic] 锚点为纯文本(R7/R8)
- U2: linkToDwtable 返回新建 columnId，前端为 FORWARD/REVERSE 锚点
  设置 data-column-id，解决同 table+record 多列锚点歧义
- U3: 重构级联删除顺序为 SELECT 收集 -> 文本恢复 -> DELETE records，
  新增 selectItemsByColumnIdWithLinkBlockId(linkBlockId IS NOT NULL)
  解耦收集条件，修复 value-clear 路径遗漏 FORWARD 恢复; per-block
  best-effort try-catch，列与记录始终删除(R4-R10)
- U5: 新增 NoteBlockContentServiceImplTest(15 场景)，扩展
  NoteColumnServiceImplTest 覆盖 AE1-AE4 与边界场景(共 48 测试通过)

+++ ruoyi-system/src/main/java/com/ruoyi/system/service/INoteBlockService.java
@@ -66,9 +66,9 @@ public interface INoteBlockService
      * 对多维表格进行双向关联
      *
      * @param noteBlockVo
-     * @return 结果
+     * @return 新建列的 columnId;失败返回 null
      */
-    int linkToDwtable(NoteBlockVo noteBlockVo);
+    Long linkToDwtable(NoteBlockVo noteBlockVo);
 
     /**
      * 对多维表格取消双向关联
